テスト用ヘルパー関数を整理し、Honoアプリ作成機能を追加。不要なファイルを削除。
